pyproject.toml
[project]
name = "metagen"
version = "0.1.0"
description = "AI/LLM-powered CLI for generating standards-compliant metadata for geospatial and tabular datasets"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.80.0",
    "click>=8.3.1",
    "langchain-community>=0.4.1",
    "langchain-litellm>=0.5.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
metagen = "metagen.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metagen"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]
